Add save/load scripts for Doodle Champion Island
This adds my current save file for the Google Doodle, "Doodle Champion
Island Games Begin!", which was that retro-inspired 16-bit RPG game that
Google made in honour of the 2021 Summer Olympics in Japan.

To use the scripts, run them in your browser on that Google Doodle page:
https://www.google.com/doodles/doodle-champion-island-games-begin

And as a backup, I've kept my most recent progress's state saved as
kitsune.json, so that I can always reload my progress even if whatever
device I was playing on ends up wiped somehow.